Refactor LogService to services package
