+++ aFWall/src/main/AndroidManifest.xml
@@ -306,7 +306,7 @@
         </receiver>
 
         <service android:name=".service.RootShellService" />
-        <service android:name=".log.LogService" />
+        <service android:name=".service.LogService" />
 
         <!-- Samsung MultiWindow Support -->
         <uses-library
