+++ aFWall/src/main/java/dev/ukanth/ufirewall/broadcast/ConnectivityChangeReceiver.java
@@ -28,7 +28,6 @@ import android.content.Intent;
 import dev.ukanth.ufirewall.Api;
 import dev.ukanth.ufirewall.InterfaceTracker;
 import dev.ukanth.ufirewall.log.Log;
-import dev.ukanth.ufirewall.log.LogService;
 import dev.ukanth.ufirewall.util.G;
 
 public class ConnectivityChangeReceiver extends BroadcastReceiver {
