+++ aFWall/src/main/java/dev/ukanth/ufirewall/preferences/PreferencesActivity.java
@@ -56,7 +56,7 @@ import dev.ukanth.ufirewall.Api;
 import dev.ukanth.ufirewall.R;
 import dev.ukanth.ufirewall.events.LogChangeEvent;
 import dev.ukanth.ufirewall.events.RulesEvent;
-import dev.ukanth.ufirewall.log.LogService;
+import dev.ukanth.ufirewall.service.LogService;
 import dev.ukanth.ufirewall.service.RootShellService;
 import dev.ukanth.ufirewall.util.G;
